requirements.txt
customtkinter==4.6.3
joblib==1.1.0
numpy==1.21.5
opencv_python==4.6.0.66
Pillow==9.3.0
PyAutoGUI==0.9.53
PyQt5==5.15.7
requests
pygetwindow
selenium